requirements.txt
# Core
flask>=3.0.0
python-dotenv>=1.0.0

# LLM Providers
langchain-google-genai>=0.1.0
langchain-core>=0.1.0
google-generativeai>=0.3.0

# NLP and Text Analysis
nltk>=3.8.0
textstat>=0.7.3
textblob>=0.17.1
vaderSentiment>=3.3.2
spacy>=3.7.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11

# Data Handling
pandas>=2.0.0

# Reporting
reportlab>=3.6.12

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0